requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
gql[all]==3.5.0
pydantic==2.5.0
redis==5.0.0
textblob==0.17.1
nltk==3.8.1
aiohttp==3.9.0
